% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_to_components.R
\name{assign_to_components}
\alias{assign_to_components}
\title{Assign variants to components}
\usage{
assign_to_components(posterior.df, cutoff = 0.8)
}
\arguments{
\item{posterior.df}{posterior.df from calc_posteriors()}

\item{cutoff}{the minimum cutoff to assign to a non-null component, default 0.8}
}
\value{
the posterior df with an additional 'component' column
}
\description{
Assign variants to components
}
