% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_compare_plot.R
\name{beta_compare_plot}
\alias{beta_compare_plot}
\title{Make a plot to compare the betas for variants assigned to different
components. Each point is one variant, each axis is the category. Error
bars are standard errors along that axis.}
\usage{
beta_compare_plot(var.df, dat, include.null = FALSE)
}
\arguments{
\item{var.df}{variant data frame from `make_var_table()`}

\item{dat}{a list with B and SE elements, variant IDs; from `prep_gwas_input()`}

\item{include.null}{whether to return the table for the null variants, defaults to false}
}
\description{
Make a plot to compare the betas for variants assigned to different
components. Each point is one variant, each axis is the category. Error
bars are standard errors along that axis.
}
