% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fdr.R
\name{calc_fdr}
\alias{calc_fdr}
\title{Calculate the false discovery rate for non-null components given the data and at
a particular posterior cutoff. Returns NA for a component if there are fewer than
five variants in it.}
\usage{
calc_fdr(posterior.df, cutoff = 0.8)
}
\arguments{
\item{posterior.df}{posterior.df from calc_posteriors()}

\item{cutoff}{the minimum cutoff to assign to a non-null component, default 0.8}
}
\value{
list of false discovery rates at that cutoff for each non-null component
}
\description{
Calculate the false discovery rate for non-null components given the data and at
a particular posterior cutoff. Returns NA for a component if there are fewer than
five variants in it.
}
