% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_heritability.R
\name{calc_heritability}
\alias{calc_heritability}
\title{Calculate the heritability for a trait.
Note: current code does not return the confidence interval;
this will be added in the next version.}
\usage{
calc_heritability(fit1, dat, posterior.df = "", cutoff = 0.8)
}
\arguments{
\item{fit1}{a model1 fit}

\item{dat}{a list with B and SE elements, variant IDs; from prep_gwas_input()}

\item{posterior.df}{optional: provide the posterior data.frame if you already
have it to avoid recalculating}

\item{cutoff}{the cutoff to assign to the non-null component, defaults to 0.8}
}
\value{
a list of the heritabilities
}
\description{
Calculate the heritability for a trait.
Note: current code does not return the confidence interval;
this will be added in the next version.
}
