% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_posteriors.R
\name{calc_posteriors}
\alias{calc_posteriors}
\title{Calculate posteriors for the variants.}
\usage{
calc_posteriors(fit, dat)
}
\arguments{
\item{fit}{the output of model1 or model2}

\item{dat}{a list with B and SE elements, variant IDs; from prep_gwas_input()}
}
\value{
a dataframe with the posteriors for each of the variants:
        for m1 this contains the variant name and only the non-null posterior,
        for m2 this contains the variant name and all four posteriors
}
\description{
Calculate posteriors for the variants.
}
