% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{.all_numeric}
\alias{.all_numeric}
\title{Check a data.frame is all numeric}
\usage{
.all_numeric(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
TRUE if all values in a data frame are numeric
}
\description{
Check a data.frame is all numeric
}
