% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_heritability.R
\name{.calc_herit_one}
\alias{.calc_herit_one}
\title{Helper function to calculate the heritability for one group (e.g. males or females)}
\usage{
.calc_herit_one(se.p2, Sigma, p, idx)
}
\arguments{
\item{se.p2}{vector of standard errors from the non-null component}

\item{Sigma}{the variance covariance matrix from an m1 fit}

\item{p}{the proportion vector from an m1 fit}

\item{idx}{the index in the list of groups}
}
\value{
the heritability of the trait in that group
}
\description{
Helper function to calculate the heritability for one group (e.g. males or females)
}
