% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_posteriors.R
\name{.calc_posterior_variant_m1}
\alias{.calc_posterior_variant_m1}
\title{Calculate the posterior probability for an m1 variant}
\usage{
.calc_posterior_variant_m1(B, SE, p, Sigma)
}
\arguments{
\item{B}{betas from the dat object}

\item{SE}{ses from the dat object}

\item{p}{proportion vector from the m1 fit}

\item{Sigma}{var-covar matrix from the m1 fit}
}
\value{
the posterior probability of belonging to the non-null component
}
\description{
Calculate the posterior probability for an m1 variant
}
