% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{.check_gwas_f_format}
\alias{.check_gwas_f_format}
\title{Check the GWAS input file to make sure that the specified columns are included.
Raises an error if the ID, B, SE columns are not present and a warning if others are
missing. Also checks that the input is bigger than 3x1.}
\usage{
.check_gwas_f_format(dat, ID, B, SE, CHR, P)
}
\arguments{
\item{dat}{the data frame resulting from reading in the gwas file}

\item{ID}{name of column with variant dds}

\item{B}{name of column with betas}

\item{SE}{name of column with ses}

\item{CHR}{name of column with chromosome number}

\item{P}{name of column name with p-vals}
}
\description{
Check the GWAS input file to make sure that the specified columns are included.
Raises an error if the ID, B, SE columns are not present and a warning if others are
missing. Also checks that the input is bigger than 3x1.
}
