% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{.check_in_dat_format}
\alias{.check_in_dat_format}
\title{Check the input data is formatted correctly}
\usage{
.check_in_dat_format(dat)
}
\arguments{
\item{dat}{input data for STAN, should be generated by `prep_gwas_input()`}
}
\description{
Check the input data is formatted correctly
}
