% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{.check_overlap_rows}
\alias{.check_overlap_rows}
\title{Check that the final list of overlapping variants is non-zero and warn
if it is small.}
\usage{
.check_overlap_rows(rows.to.keep)
}
\arguments{
\item{rows.to.keep}{list of variants after filtering}
}
\description{
Check that the final list of overlapping variants is non-zero and warn
if it is small.
}
