% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{.check_posterior_df}
\alias{.check_posterior_df}
\title{Check a posterior df is formatted correctly}
\usage{
.check_posterior_df(posterior.df)
}
\arguments{
\item{posterior.df}{a data frame with posteriors}
}
\description{
Check a posterior df is formatted correctly
}
