% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{.check_read_in_params}
\alias{.check_read_in_params}
\title{Check the input parameters for `prep_gwas_input()`}
\usage{
.check_read_in_params(gwas.files, cat.labels, se.cut)
}
\arguments{
\item{gwas.files}{list of files containing GWAS input}

\item{cat.labels}{vector of labels for the files specifying their categories}

\item{se.cut}{the max standard error allowed}
}
\description{
Check the input parameters for `prep_gwas_input()`
}
