% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{.check_var_to_keep}
\alias{.check_var_to_keep}
\title{Check that the variants to keep ids overlap with the data ids.}
\usage{
.check_var_to_keep(var.keep, data.ids)
}
\arguments{
\item{var.keep}{list of variants to keep}

\item{data.ids}{list of variant ids from the data}
}
\description{
Check that the variants to keep ids overlap with the data ids.
}
