% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_gwas_input.R
\name{.extract_overlapping_rows}
\alias{.extract_overlapping_rows}
\title{Helper function to extract overlapping rows from a list of datasets.
This finds the rows present in all datasets and reorders all datasets
to have the same order.}
\usage{
.extract_overlapping_rows(list.ds)
}
\arguments{
\item{list.ds}{a list of datasets with betas and ses}
}
\value{
the reordered and filtered set of datasets
}
\description{
Helper function to extract overlapping rows from a list of datasets.
This finds the rows present in all datasets and reorders all datasets
to have the same order.
}
