% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{.my_assert}
\alias{.my_assert}
\title{Custom assertion function, prints formatted error messages when they fail.}
\usage{
.my_assert(error_msg, test)
}
\arguments{
\item{error_msg}{error message to print on failure}

\item{test}{the value to test}
}
\description{
Prints an error message if the assertion fails, otherwise does nothing.
Error message is of the format (and does not include the function call):
   "Error: <error_msg>"
}
