% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{.my_warn}
\alias{.my_warn}
\title{Custom warning function, prints formatted warning messages when they fail.}
\usage{
.my_warn(warn_msg, test)
}
\arguments{
\item{warn_msg}{warning message to print on failure}

\item{test}{the value to test}
}
\description{
Prints an warning message if the assertion fails, otherwise does nothing.
Warning message is of the format (and does not include the function call):
   "Warning: <warn_msg>"
}
