% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_gwas_input.R
\name{.prep_dat_stan}
\alias{.prep_dat_stan}
\title{Helper function to prep the dataset for stan}
\usage{
.prep_dat_stan(list.ds)
}
\arguments{
\item{list.ds}{a list of datasets, already filtered so the rows are the same}
}
\value{
the data put together such that there is one column per dataset
}
\description{
Helper function to prep the dataset for stan
}
