% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fit_params.R
\name{get_gen_cor}
\alias{get_gen_cor}
\title{Get the genetic correlation associated with a model1 fit}
\usage{
get_gen_cor(fit1, ndim = 2)
}
\arguments{
\item{fit1}{fitted rstan model from model1 fit}

\item{ndim}{optional parameter to specify number of categories (e.g. male, female), defaults to 2}
}
\value{
list of pairwise genetic correlations ndim-1 long (for ndim==2 it is a single value)
}
\description{
Get the genetic correlation associated with a model1 fit
}
