% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fit_params.R
\name{get_gen_cor_ci}
\alias{get_gen_cor_ci}
\title{Get a ninety-five percent confidence interval for the genetic correlation}
\usage{
get_gen_cor_ci(fit1, ndim = 2)
}
\arguments{
\item{fit1}{fitted rstan model from model1 fit}

\item{ndim}{optional parameter to specify number of categories (e.g. male, female), defaults to 2}
}
\value{
lists of lower (l) and upper (u) CIs for the each of the pairwise genetic correlations
       each sublist will be ndim-1 long (for ndim==2 it is a single value), ordered by the beta pairs
}
\description{
Get a ninety-five percent confidence interval for the genetic correlation
}
