% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fit_params.R
\name{get_proportions}
\alias{get_proportions}
\title{Get the proportions associated with a fit}
\usage{
get_proportions(fit)
}
\arguments{
\item{fit}{fitted rstan model from model1 or model2 fit}
}
\value{
vector containing the proportions for the fit
}
\description{
Get the proportions associated with a fit
}
