% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fit_params.R
\name{get_var_covar_matrix}
\alias{get_var_covar_matrix}
\title{Get the variance-covariance matrix associated with a model1 fit}
\usage{
get_var_covar_matrix(fit1, ndim = 2)
}
\arguments{
\item{fit1}{fitted rstan model from model1 fit}

\item{ndim}{optional parameter to specify number of categories (e.g. male, female), defaults to 2}
}
\value{
variance covariance matrix of size ndim x ndim
}
\description{
Get the variance-covariance matrix associated with a model1 fit
}
