% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_var_table.R
\name{make_var_table}
\alias{make_var_table}
\title{Make a variant table to easily view the output of m2}
\usage{
make_var_table(dat, posterior.df, cutoff = 0.8, include.null = FALSE)
}
\arguments{
\item{dat}{a list with B and SE elements, variant IDs; from prep_gwas_input()}

\item{posterior.df}{posterior.df from calc_posteriors()}

\item{cutoff}{the minimum cutoff to assign to a non-null component, default 0.8}

\item{include.null}{whether to return the table for the null variants, defaults to false}
}
\value{
a table specifying the variants, their betas, pvals, and posteriors
}
\description{
Make a variant table to easily view the output of m2
}
