% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model1_stan.R
\name{model1_stan}
\alias{model1_stan}
\title{Estimate genetic parameters using SEMM.}
\usage{
model1_stan(dat = NULL, B = NULL, SE = NULL, ...)
}
\arguments{
\item{dat}{input data object, produced from prep_gwas_input()}

\item{B}{optional, if dat is not provided,
a data frame containing beta estimates from GWAS summary statistics.
each row corresponds to a variant and each column corresponds to
a group (e.g. males and females)}

\item{SE}{optional, if dat is not provided,
a data frame containing SE estimates from GWAS summary statistics,
 each row/column refers to the same variant and group as the beta values}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling` fitted to model1
}
\description{
This function takes as input paired betas and standard errors from GWAS summary
statistics and outputs a fitted model that contains genetic correlation
and heritability estimates.
}
