% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2alt_stan.R
\name{model2alt_stan}
\alias{model2alt_stan}
\title{Fit alternate (no sex-specific variants) model to the data. This is used
for comparison with model2 to test for evidence for sex-specific components.}
\usage{
model2alt_stan(dat = NULL, B = NULL, SE = NULL, ...)
}
\arguments{
\item{dat}{input data object, produced from prep_gwas_input()}

\item{B}{optional, if dat is not provided,
a data frame containing beta estimates from GWAS summary statistics.
each row corresponds to a variant and each column corresponds to
a group (e.g. males and females)}

\item{SE}{optional, if dat is not provided,
a data frame containing SE estimates from GWAS summary statistics,
 each row/column refers to the same variant and group as the beta values}
}
\value{
An object of class `stanfit` returned by `rstan::sampling` fitted to model2
}
\description{
Fit alternate (no sex-specific variants) model to the data. This is used
for comparison with model2 to test for evidence for sex-specific components.
}
