% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2loglik_stan.R
\name{model2loglik_stan}
\alias{model2loglik_stan}
\title{Identify sex-specific variants using SEMM.}
\usage{
model2loglik_stan(dat = NULL, B = NULL, SE = NULL, ...)
}
\arguments{
\item{dat}{input data object, produced from prep_gwas_input()}

\item{B}{optional, if dat is not provided,
a data frame containing beta estimates from GWAS summary statistics.
each row corresponds to a variant and each column corresponds to
a group (e.g. males and females)}

\item{SE}{optional, if dat is not provided,
a data frame containing SE estimates from GWAS summary statistics,
 each row/column refers to the same variant and group as the beta values}
}
\value{
An object of class `stanfit` returned by `rstan::sampling` fitted to model2
}
\description{
This function takes as input paired betas and standard errors from GWAS summary
statistics and outputs a fitted four component model that can be used to assign
variants to components.
}
\details{
This runs a variant of the model2 code that calculates the loglikelihood - as a result
it runs much more slowly. This is needed for comparison with the alternate model.
}
