% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_gwas_input.R
\name{prep_gwas_input}
\alias{prep_gwas_input}
\title{Reads in multiple GWAS files and reformats to a data object that can be used for model fitting.}
\usage{
prep_gwas_input(gwas.files, f.labels = "", se.cut = 0.2,
  var.keep = "", BETA = "BETA", ID = "ID", CHR = "CHR",
  SE = "SE", P = "P")
}
\arguments{
\item{gwas.files}{a list of locations of files containing GWAS summary statistics,
      this program uses the "ID", "CHR", BETA","SE", and "P" columns for the analysis.
      Please rename your columns to match this or specify this as parameters.

---- optional params ----}

\item{f.labels}{optionally provide a vector of labels for the files (e.g. c("female", "male")),
defaults to cat1, ..., catn}

\item{se.cut}{optionally specify the max standard error allowed, defaults to 0.2}

\item{var.keep}{optional list of variants to keep, often used for LD or other filtering,
defaults to keeping all}

\item{BETA}{optional column name with betas, default "BETA"}

\item{ID}{optional column name with variant ids, default "ID"}

\item{CHR}{optional column name with chromosome number, default "CHR"}

\item{SE}{optional column name with ses, default "SE"}

\item{P}{optional column name with p-vals, default "P"
------------------------}
}
\value{
a data object containing betas and standard errors formatted
       for model fitting, also contains variant ids
}
\description{
Reads in multiple GWAS files and reformats to a data object that can be used for model fitting.
}
