% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_m1_data.R
\name{sim_ses}
\alias{sim_ses}
\title{Simulate standard errors by sampling from a half-normal distribution.}
\usage{
sim_ses(N, max.se = 0.4, min.se = 0.005)
}
\arguments{
\item{N}{the number of simulated standard errors to generate}
}
\value{
a vector of N squared simulated "standard errors"
}
\description{
Parameters and filtering were selected to approximate the true distribution.
}
