% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fit_params.R
\name{get_vars}
\alias{get_vars}
\title{Get the variances from the m2 fit}
\usage{
get_vars(fit2)
}
\arguments{
\item{fit2}{the output of an m2}
}
\value{
a list of the variances estimated from the m2 fit
}
\description{
Get the variances from the m2 fit
}
